pyproject.toml
[project]
name = "inventory-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.3",
    "bcrypt>=5.0.0",
    "black>=26.1.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.128.6",
    "flake8>=7.3.0",
    "greenlet>=3.3.1",
    "mypy>=1.19.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "pymysql>=1.1.2",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.19.0",
]
